+++ src/client/modules/wallet/ConnectionManager.tsx
