+++ src/client/modules/wallet/ethereum-logo.png
