+++ src/client/modules/wallet/index.ts
@@ -10,6 +10,7 @@
 
 import { useWeb3React } from "@web3-react/core";
 import { CHAIN_INFO } from "./chains";
+import getLibrary from "./getLibrary";
 
 import { useConnect, useEagerConnect } from "./hooks";
 import useChain from "./useChain";
@@ -21,5 +22,6 @@ export {
   useEagerConnect,
   useEnsName,
   useWeb3React,
+  getLibrary,
   CHAIN_INFO,
 };
