--- src/modules/nfts/index.ts
@@ -1,3 +1,0 @@
-import { useNfts } from "./rio";
-
-export { useNfts };
