+++ src/pages/_app.tsx
@@ -1,8 +1,8 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
 import { Web3ReactProvider } from "@web3-react/core";
-import { QueryClient, QueryClientProvider, useQuery } from "react-query";
-import getLibrary from "../modules/wallet/getLibrary";
+import { QueryClient, QueryClientProvider } from "react-query";
+import { getLibrary } from "client/modules/wallet";
 
 const queryClient = new QueryClient();
